fix(api): 修复分页接口返回值类型错误
- 统一将分页参数 pageNum 和 pageSize 强制转换为 long 类型
- 修正 AppInfoController、ConfigController、DictController、DictItemController、LanguageController 和 VariableController 中分页响应格式
- 防止因类型不匹配导致的接口调用异常，提高系统稳定性

+++ remi-system/remi-system-web/src/main/java/com/remisoft/system/web/controller/AppInfoController.java
@@ -86,7 +86,7 @@ public class AppInfoController {
             @Parameter(description = "应用名称模糊搜索") @RequestParam(required = false) String appName,
             @Parameter(description = "状态") @RequestParam(required = false) String status) {
         IPage<AppInfoVO> page = service.page(pageNum, pageSize, appName, status);
-        return BaseResponse.successPage(page.getTotal(), pageNum, pageSize, page.getRecords());
+        return BaseResponse.successPage(page.getTotal(), (long) pageNum, (long) pageSize, page.getRecords());
     }
 
     /**
